fix(helper): keep storage stat arithmetic portable

+++ crates/helper-store/src/lib.rs
@@ -826,9 +826,9 @@ impl Store {
                     return None;
                 }
                 let stat = unsafe { stat.assume_init() };
-                let block_size = stat.f_frsize.max(1) as u64;
-                let total = (stat.f_blocks as u64).saturating_mul(block_size);
-                let available = (stat.f_bavail as u64).saturating_mul(block_size);
+                let block_size = stat.f_frsize.max(1);
+                let total = stat.f_blocks.saturating_mul(block_size);
+                let available = stat.f_bavail.saturating_mul(block_size);
                 Some((total.saturating_sub(available), available, total))
             })
         };
